Uploading and cleaning Bostong Housing Project
type: chore

This is one of the project for Udacity's Machine Learning Nanodegree
to help students solidify their skills from the course material. It goes
through simple feature extraction from a dataset, data exploration,
model building and validating, and predicting home prices based on
a certain set of known parameters.

The code is wrtting in Python v2.7 on a Jupyter Notebook for easy
reproducibility and visualization. A handful of files were removed
from the original submission of the project to keep the user focused
on only the files needed.
-housing.csv is the data used in the project for our analysis
-project_description.md is Udacity's markdown file and summary of
the project and what will be used for our GitHub repository
-visuals.py is a script that simplifies the visualiations of the models
we train and test in the project. This helps give us insight to model
performance on different permutations of the GridSearchCV
-boston_houstin.ipynb is the Jupyter Notebook with the full code
and models sent in for submission